+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added new addon to check container optimization with `dive`
 * Improved conftest.check() function. It tells the user how to create comparison files if the file not found exception.
 * Improved `conftest.check()` function. Now AssertionError is handled, the user will get receive how to update comparison files.
+* Added new class `PredefinedImages` in `gcip.addons.container`. Allows access to container images, that are widley used.
 
 ### Changed
 
@@ -31,6 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   a dummy string is returned instead of raising a KeyError.
 * The `gcip.Pipeline` has now an `add_services()` method instead `add_service()` allowing to pass multiple
   services at once.
+* Now jobs with hardcoded images, now using PredefinedImages images instead.
+
 
 ## [0.5.0] - 2021-03-16